requirements.txt
chess == 1.10.0
psutil == 5.9.5
PyYAML == 6.0.1
requests == 2.31.0
tenacity == 8.2.2
openai == 0.27.2
chatchess == 1.1.7
